QoL & Performance updates
Name changes:
placeholder -> master #more descriptive

Added:
imagewindowgeometry as an initializer to pass geometry, because we want to avoid using
update_idletasks(), and because master.winfo_height() == 1 when we don't use update_idletasks().

Changes:
Enables zooming when cursor is not above the image.
Made the code a bit more readable by adding spaces and comments.

Canvas is now set to width and height of master window. Reason is below.
The __imframe widget's weight is now 0. This together with the canvas change will remove scrollbars.

I removed the scrollbars because panning and zooming are fully functional without them visible, and
with them gone, it doesn't break image centering and resizing (to be implemented)